deploy task optimized
