--- evmv2/tasks/swapIn.js
@@ -1,35 +1,0 @@
-
-function stringToHex(str) {
-    return str.split("").map(function(c) {
-        return ("0" + c.charCodeAt(0).toString(16)).slice(-2);
-    }).join("");
-}
-
-const maticReceiptProof = '0x0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000012000000000000000000000000000000000000000000000000000000000000006a0180f22269bce024a7a4100ff237751ff3d19ff3dc085ca2af53f54d48e6612c7135b845832cb522cd103b85a86dee9f5df051137e081b19de421eff1bfdd71ea2c47ea0dbf24d7777c5904e9855202a0adcd78ce1c26dfb3a0c808a000ecf184302b39841515efb2e338cbe12bf6e860fb2f73fe5f25fbe9bf946483933df8e900000000000000000000000000000000000000000000000000000000000009e0000000000000000000000000000000000000000000000000000000000000116000000000000000000000000000000000000000000000000000000000000011a000000000000000000000000000000000000000000000000000000000000001c0000000000000000000000000096bf1097f3af73b716eab545001d97b2cf1fb2000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000240000000000000000000000000000000000000000000000000000000000000028000000000000000000000000000000000000000000000000000000000000002c0000000000000000000000000000000000000000000000000000000000008688400000000000000000000000000000000000000000000000000000000007a12000000000000000000000000000000000000000000000000000000000000023e94000000000000000000000000000000000000000000000000000000006391a52d00000000000000000000000000000000000000000000000000000000000003e000000000000000000000000000000000000000000000000000000000000005000000000000000000000000000000000000000000000000000000000000000540000000000000000000000000000000000000000000000000000000174876e8000000000000000000000000000000000000000000000000000000000000000020eef5bb8f08ae6dce2c23be3a63c8714bc2a233522f28e474461a61863bd9f7e20000000000000000000000000000000000000000000000000000000000000020fa19969a562a8b917bfac6b1994ee1c80a85610bce9401cd7a3b0b1513d7edbc0000000000000000000000000000000000000000000000000000000000000020edb832983b3a6f1d7c93f377fee8fb72946c60f080827b8995f880deff82e741000000000000000000000000000000000000000000000000000000000000002005a8fbb1207725edf878eb6831289589e2dd234a30b0d4a934e637314256a71d00000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000100000000000200000000000000000000000000000000000000020000000000000000000000004000000000000020000800010000000000800000000800000000002080000000800000000000001000000000100002000000000000000800080000000000000000000000001000000000000000000000000000000000000000000000000000004000000000000000000002000000000000000000000000000000000020000002000040400000000000000000001200000000000000000000000000000000000000000000000000002000001000000000000000000000000000000000000000000000400080000000000000000000000000000000000000000000000000000000000000000000000000f5d883010000846765746888676f312e31352e36856c696e757800000000000000f8d3c0c0c080b8412b74a146893ff4f80aa5105516402a23872eaf8d4dd4446d2607f9de7b6ef564797aa9b941af71fcf2de4a809de35c3717ad6c33fb0f95ef407f167da882ebd201f8440bb8400f0d1d411e5ba2ebef802a0520cb9eed75198b51787132185e2f3ec53df2e0b02cbe2ec334c4e45baa81df80d690fa56bcc4024b607b87fa7d8e085ea330deb380f8440fb8400b5b1e56d00ad5baf52f671be78b8f41c923945d8cb6c140bd6691f6cc672414176577100b9bdc652ff2bbbd20312d89c32464b6bdab633d807366f30d3ff805800000000000000000000000000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000e0000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001200000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000014000000000000000000000000000000000000000000000000000000000000001c0000000000000000000000000000000000000000000000000000000000000028000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000412b74a146893ff4f80aa5105516402a23872eaf8d4dd4446d2607f9de7b6ef564797aa9b941af71fcf2de4a809de35c3717ad6c33fb0f95ef407f167da882ebd20100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000b0000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000400f0d1d411e5ba2ebef802a0520cb9eed75198b51787132185e2f3ec53df2e0b02cbe2ec334c4e45baa81df80d690fa56bcc4024b607b87fa7d8e085ea330deb3000000000000000000000000000000000000000000000000000000000000000f0000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000400b5b1e56d00ad5baf52f671be78b8f41c923945d8cb6c140bd6691f6cc672414176577100b9bdc652ff2bbbd20312d89c32464b6bdab633d807366f30d3ff805000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000000705f907020183023e94b9010000000000000000000000000000000001000000000002000000000000000000000000000000000000000200000000000000000000000040000000000000200008000100000000008000000008000000000020800000008000000000000010000000001000020000000000000008000800000000000000000000000010000000000000000000000000000000000000000000000000000040000000000000000000020000000000000000000000000000000000200000020000404000000000000000000012000000000000000000000000000000000000000000000000000020000010000000000000000000000000000000000000000000004000800000000000f905f7f89b945f91a91dba041073858e1e2236605c4db2f5488cf863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa00000000000000000000000008c9b3caf7dedd3003f53312779c1b92ba1625d94a000000000000000000000000074116d39a9e4ed0d160999f042f69286b5bcf1a3a000000000000000000000000000000000000000000000000014d1120d7b160000f89b945f91a91dba041073858e1e2236605c4db2f5488cf863a08c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925a00000000000000000000000008c9b3caf7dedd3003f53312779c1b92ba1625d94a000000000000000000000000074116d39a9e4ed0d160999f042f69286b5bcf1a3a00000000000000000000000000000000000000000000000000000000000000000f89b945f91a91dba041073858e1e2236605c4db2f5488cf863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa000000000000000000000000074116d39a9e4ed0d160999f042f69286b5bcf1a3a00000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000136dcc951d8c0000f9041d9474116d39a9e4ed0d160999f042f69286b5bcf1a3f863a0ca1cf8cebf88499429cca8f87cbca15ab8dafd06702259a5344ddce89ef3f3a5a000000000000000000000000000000000000000000000000000000000000000d4a00000000000000000000000000000000000000000000000000000000000013881b903a06f42a4a333325ff2ade47b53c8dcdc982720714edbb60ef4791a6e387f82e2d400000000000000000000000000000000000000000000000000000000000000c000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000140000000000000000000000000000000000000000000000000136dcc951d8c0000000000000000000000000000000000000000000000000000000000000000018000000000000000000000000000000000000000000000000000000000000000141e01cf4503808fb30f17806035a87cf5a521772700000000000000000000000000000000000000000000000000000000000000000000000000000000000000148c9b3caf7dedd3003f53312779c1b92ba1625d9400000000000000000000000000000000000000000000000000000000000000000000000000000000000000148c9b3caf7dedd3003f53312779c1b92ba1625d940000000000000000000000000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000001c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000de0b6b3a76400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000020000000000000000000000001e01cf4503808fb30f17806035a87cf5a5217727000000000000000000000000e1d8eab4e616156e11e1c59d1a0e0efed66f4cfa0000000000000000000000000000000000000000000000000000000000000014e1d8eab4e616156e11e1c59d1a0e0efed66f4cfa0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000070ef9070b822080b90705f907020183023e94b9010000000000000000000000000000000001000000000002000000000000000000000000000000000000000200000000000000000000000040000000000000200008000100000000008000000008000000000020800000008000000000000010000000001000020000000000000008000800000000000000000000000010000000000000000000000000000000000000000000000000000040000000000000000000020000000000000000000000000000000000200000020000404000000000000000000012000000000000000000000000000000000000000000000000000020000010000000000000000000000000000000000000000000004000800000000000f905f7f89b945f91a91dba041073858e1e2236605c4db2f5488cf863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa00000000000000000000000008c9b3caf7dedd3003f53312779c1b92ba1625d94a000000000000000000000000074116d39a9e4ed0d160999f042f69286b5bcf1a3a000000000000000000000000000000000000000000000000014d1120d7b160000f89b945f91a91dba041073858e1e2236605c4db2f5488cf863a08c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925a00000000000000000000000008c9b3caf7dedd3003f53312779c1b92ba1625d94a000000000000000000000000074116d39a9e4ed0d160999f042f69286b5bcf1a3a00000000000000000000000000000000000000000000000000000000000000000f89b945f91a91dba041073858e1e2236605c4db2f5488cf863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa000000000000000000000000074116d39a9e4ed0d160999f042f69286b5bcf1a3a00000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000136dcc951d8c0000f9041d9474116d39a9e4ed0d160999f042f69286b5bcf1a3f863a0ca1cf8cebf88499429cca8f87cbca15ab8dafd06702259a5344ddce89ef3f3a5a000000000000000000000000000000000000000000000000000000000000000d4a00000000000000000000000000000000000000000000000000000000000013881b903a06f42a4a333325ff2ade47b53c8dcdc982720714edbb60ef4791a6e387f82e2d400000000000000000000000000000000000000000000000000000000000000c000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000140000000000000000000000000000000000000000000000000136dcc951d8c0000000000000000000000000000000000000000000000000000000000000000018000000000000000000000000000000000000000000000000000000000000000141e01cf4503808fb30f17806035a87cf5a521772700000000000000000000000000000000000000000000000000000000000000000000000000000000000000148c9b3caf7dedd3003f53312779c1b92ba1625d9400000000000000000000000000000000000000000000000000000000000000000000000000000000000000148c9b3caf7dedd3003f53312779c1b92ba1625d940000000000000000000000000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000001c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000de0b6b3a76400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000020000000000000000000000001e01cf4503808fb30f17806035a87cf5a5217727000000000000000000000000e1d8eab4e616156e11e1c59d1a0e0efed66f4cfa0000000000000000000000000000000000000000000000000000000000000014e1d8eab4e616156e11e1c59d1a0e0efed66f4cfa000000000000000000000000000000000000000000000000000000000000'
-const bscReceiptProof = '0x0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000012000000000000000000000000000000000000000000000000000000000000006a025fb3b1d6f045ed6bf02c89501cda0a3e1ba4fdd0c3ac62666190c27adab49e1061889591a3de36d5880e2a59d52a619c7e77b69337ff7e7c8303fb60df4d9c32651dec7274460f81c4f15883eee2e5ac2a323fecf8c2073840a42fd00b2e382237093dfbabc3391fb44592c154ac894118aef6619407c2eb1146c4584b2a4dd00000000000000000000000000000000000000000000000000000000000009e0000000000000000000000000000000000000000000000000000000000000116000000000000000000000000000000000000000000000000000000000000011a000000000000000000000000000000000000000000000000000000000000001c00000000000000000000000002f3079a1c1c0995a1c9803853d1b8444cce0aa9f00000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000240000000000000000000000000000000000000000000000000000000000000028000000000000000000000000000000000000000000000000000000000000002c0000000000000000000000000000000000000000000000000000000000008689e00000000000000000000000000000000000000000000000000000000007a12000000000000000000000000000000000000000000000000000000000000023e7c000000000000000000000000000000000000000000000000000000006391a5af00000000000000000000000000000000000000000000000000000000000003e000000000000000000000000000000000000000000000000000000000000005000000000000000000000000000000000000000000000000000000000000000540000000000000000000000000000000000000000000000000000000174876e8000000000000000000000000000000000000000000000000000000000000000020e485274177db55f188ce93827eac2410562f58107b5a6d4e0d08ea9b5104626f00000000000000000000000000000000000000000000000000000000000000207bea1b7dc5d1f0c81b7085a44b4511b791197ab752a6330efe5889118585e4d90000000000000000000000000000000000000000000000000000000000000020048d29e4b0568b8a104b696d4a3272c793095d69d687eea508b9ba85dc89966c000000000000000000000000000000000000000000000000000000000000002006be21a78686bf4aea559f36beb2244a2b6d97a4258cbb8d1786b84050038a2400000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000100000000200200000000000000000000000000000000000000020000000000000000000000004000000000000020000800010000000000000000000800000000002000000000800000000000001000000000000002000000000000000800080000000000000000000000001000000000000000000000000000000000000000000000000000004000000000000000000002000000000000000000000000000000000020000002000040400000000000000000001200000000000000000000000400000000000000000000000000002000001010000000000000000000000000000000000000000000400080000000000000000000000000000000000000000000000000000000000000000000000000f5d883010000846765746888676f312e31352e36856c696e757800000000000000f8d3c0c0c080b841636f16493c80d6a4f0de3e175df49165d4b09dc5b71a986b5bb9ad0a1adbc5de10e3eeffaa01e442655427f4e97e08d3f4fef70ce1fe3f3d56865052edd71cb501f8440eb840113f8109121b52a1632efce1441b77c4771f7c119a47d6e8e43cbd91a1633ed7024cd77b15b54f990ccf233a32a13cdacfdc1134aa70700ea1d3dc81905d64d180f8440fb84004f8efce4cec8ee890c077ec96eb4dddc17cc00b585869a34a10fef99a5375680608ed95c9c1bdabb6ab75c0fd43b887dd9f880811b5b9270c3b3d549c3610b6800000000000000000000000000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000e0000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001200000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000014000000000000000000000000000000000000000000000000000000000000001c000000000000000000000000000000000000000000000000000000000000002800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041636f16493c80d6a4f0de3e175df49165d4b09dc5b71a986b5bb9ad0a1adbc5de10e3eeffaa01e442655427f4e97e08d3f4fef70ce1fe3f3d56865052edd71cb50100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000e000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000040113f8109121b52a1632efce1441b77c4771f7c119a47d6e8e43cbd91a1633ed7024cd77b15b54f990ccf233a32a13cdacfdc1134aa70700ea1d3dc81905d64d1000000000000000000000000000000000000000000000000000000000000000f00000000000000000000000000000000000000000000000000000000000000600000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004004f8efce4cec8ee890c077ec96eb4dddc17cc00b585869a34a10fef99a5375680608ed95c9c1bdabb6ab75c0fd43b887dd9f880811b5b9270c3b3d549c3610b6000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000000705f907020183023e7cb9010000000000000000000000000000000001000000002002000000000000000000000000000000000000000200000000000000000000000040000000000000200008000100000000000000000008000000000020000000008000000000000010000000000000020000000000000008000800000000000000000000000010000000000000000000000000000000000000000000000000000040000000000000000000020000000000000000000000000000000000200000020000404000000000000000000012000000000000000000000004000000000000000000000000000020000010100000000000000000000000000000000000000000004000800000000000f905f7f89b945f91a91dba041073858e1e2236605c4db2f5488cf863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa00000000000000000000000008c9b3caf7dedd3003f53312779c1b92ba1625d94a000000000000000000000000074116d39a9e4ed0d160999f042f69286b5bcf1a3a000000000000000000000000000000000000000000000000014d1120d7b160000f89b945f91a91dba041073858e1e2236605c4db2f5488cf863a08c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925a00000000000000000000000008c9b3caf7dedd3003f53312779c1b92ba1625d94a000000000000000000000000074116d39a9e4ed0d160999f042f69286b5bcf1a3a00000000000000000000000000000000000000000000000000000000000000000f89b945f91a91dba041073858e1e2236605c4db2f5488cf863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa000000000000000000000000074116d39a9e4ed0d160999f042f69286b5bcf1a3a00000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000136dcc951d8c0000f9041d9474116d39a9e4ed0d160999f042f69286b5bcf1a3f863a0ca1cf8cebf88499429cca8f87cbca15ab8dafd06702259a5344ddce89ef3f3a5a000000000000000000000000000000000000000000000000000000000000000d4a00000000000000000000000000000000000000000000000000000000000000061b903a00315b59acd5e55d73985425186b1cb203aad632c62550c0e229121c076bc264f00000000000000000000000000000000000000000000000000000000000000c000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000140000000000000000000000000000000000000000000000000136dcc951d8c0000000000000000000000000000000000000000000000000000000000000000018000000000000000000000000000000000000000000000000000000000000000143f1e91bfc874625f4ee6ef6d8668e7929188237300000000000000000000000000000000000000000000000000000000000000000000000000000000000000148c9b3caf7dedd3003f53312779c1b92ba1625d9400000000000000000000000000000000000000000000000000000000000000000000000000000000000000148c9b3caf7dedd3003f53312779c1b92ba1625d940000000000000000000000000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000001c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000de0b6b3a76400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000020000000000000000000000003f1e91bfc874625f4ee6ef6d8668e79291882373000000000000000000000000593f6f6748dc203dfa636c299eea6a39c0734eed0000000000000000000000000000000000000000000000000000000000000014593f6f6748dc203dfa636c299eea6a39c0734eed0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000070ef9070b822080b90705f907020183023e7cb9010000000000000000000000000000000001000000002002000000000000000000000000000000000000000200000000000000000000000040000000000000200008000100000000000000000008000000000020000000008000000000000010000000000000020000000000000008000800000000000000000000000010000000000000000000000000000000000000000000000000000040000000000000000000020000000000000000000000000000000000200000020000404000000000000000000012000000000000000000000004000000000000000000000000000020000010100000000000000000000000000000000000000000004000800000000000f905f7f89b945f91a91dba041073858e1e2236605c4db2f5488cf863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa00000000000000000000000008c9b3caf7dedd3003f53312779c1b92ba1625d94a000000000000000000000000074116d39a9e4ed0d160999f042f69286b5bcf1a3a000000000000000000000000000000000000000000000000014d1120d7b160000f89b945f91a91dba041073858e1e2236605c4db2f5488cf863a08c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925a00000000000000000000000008c9b3caf7dedd3003f53312779c1b92ba1625d94a000000000000000000000000074116d39a9e4ed0d160999f042f69286b5bcf1a3a00000000000000000000000000000000000000000000000000000000000000000f89b945f91a91dba041073858e1e2236605c4db2f5488cf863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa000000000000000000000000074116d39a9e4ed0d160999f042f69286b5bcf1a3a00000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000136dcc951d8c0000f9041d9474116d39a9e4ed0d160999f042f69286b5bcf1a3f863a0ca1cf8cebf88499429cca8f87cbca15ab8dafd06702259a5344ddce89ef3f3a5a000000000000000000000000000000000000000000000000000000000000000d4a00000000000000000000000000000000000000000000000000000000000000061b903a00315b59acd5e55d73985425186b1cb203aad632c62550c0e229121c076bc264f00000000000000000000000000000000000000000000000000000000000000c000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000140000000000000000000000000000000000000000000000000136dcc951d8c0000000000000000000000000000000000000000000000000000000000000000018000000000000000000000000000000000000000000000000000000000000000143f1e91bfc874625f4ee6ef6d8668e7929188237300000000000000000000000000000000000000000000000000000000000000000000000000000000000000148c9b3caf7dedd3003f53312779c1b92ba1625d9400000000000000000000000000000000000000000000000000000000000000000000000000000000000000148c9b3caf7dedd3003f53312779c1b92ba1625d940000000000000000000000000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000001c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000de0b6b3a76400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000020000000000000000000000003f1e91bfc874625f4ee6ef6d8668e79291882373000000000000000000000000593f6f6748dc203dfa636c299eea6a39c0734eed0000000000000000000000000000000000000000000000000000000000000014593f6f6748dc203dfa636c299eea6a39c0734eed000000000000000000000000000000000000000000000000000000000000'
-const near2polygonProof = '0xd33a28a20000000000000000000000000000000000000000000000004d41500100000001000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000024800000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000154000000000000000000000000000000000000000000000000000000000000014c54c44ede2e54ae363e2c355aff1141cb0e03ceef6dce5ee9a981e92bbbfb9edf0e19fbf8b0448d8b2d50cd386d9a8fde388caa0e9ea914cfdcbeb58f6f3cc9ee5ea66780600000000a62c1450d662861991aff6bd35555ad3854fc937476e411a1f1ea3a25ce259664ef4d14151cde795dfbc6eeda9998f3f4a51bbf46cabbdf1e58ceb4a4bbacdefca984ec465a805fec9e3d4f95af5345593decc454257cedbb2889d11fe07533e001d36b7cccd72d2b44e81c39faec73047eb5e5238a2373fa5effccf48af1e8081279fb9475f3017f34b364e97d53023bc0c099d2d7c8fd7d400f9811516a034696dc081099ec36ebd2c32da10cca95246a46bf042598c12196eeb3797e4dacf7559e738c008191cb032a597b540d318d6d319d40b0e3ae43d254b12c17fb6c34de4cc7eb2ec30d2012b00000000050000006e6f64653200e9f1e1f48f966f3cb086e238c8f2d9d6a54063a8e757cf5587bf429a5d2292c2fa69d3b65f8bd11a7156be64b101000000050000006e6f646531004d7a80636b24694ba9c928394118cd539bf863f5e5773a0de27c7677d3eb0a87c329889ef3dbc20a34fb0a4db101000000050000006e6f646533000e8202a3ff22a3df1bf2221f5079005e0bc1438437a2811d513e795d797a58243962821b02e27b06f3aecf36b101000000050000006e6f646530005edad8e09c8a0569c66fcee79007935e65f5bab6df5bff201e71d7d0efe83472ea2ef88fe96949ccb93805b0b001000000160000006175726f72612e706f6f6c2e663836333937332e6d30007fdc7a529b631deb0c6c75e1893696781b3ff4d3e4a8fadee70dce85b80113c866901edd6efbddf377aa1eefb3000000001b00000063727970746f676172696b2e706f6f6c2e663836333937332e6d3000de6c9696f96a93e27bdb465b90eaf00718fae027c339202c379cd18b9f354a18cad2a0c801508e6b4621eb8caf000000001600000030316e6f64652e706f6f6c2e663836333937332e6d30002850e20b8ce610aee3d913441ff7ade411d5524f445dfd8c2fc7784d0e5155ad24b18f7eb3058fe13e0abd3a8000000000170000006c6567656e64732e706f6f6c2e663836333937332e6d30009012890b00163561ec8bd99da8715c61c32b5be92f3dc00acd7cd9479f319298eb36eebbd91ecf49a0ec60d5680000000019000000657665727374616b652e706f6f6c2e663836333937332e6d3000317f139e9df2e803cee46652ac7701aec09154c3f8cfd13acb24180069066fa2a69df3f3755e29accbae166364000000001900000063686f7275736f6e652e706f6f6c2e663836333937332e6d30002491a39951692ea9824a9821c6ffadb15aed6801ebbb587d225d4bb404110dec97a77938b9cdd736bef46cb35d00000000170000006e6f64656173792e706f6f6c2e663836333937332e6d30000ff05f1e15d447bfd0fc36be2416b585b3c4238cdfc644d3da84ef09e1ea2d4a49acbfa3dc73bf718d90b0b15c000000001a0000007374616b656c795f76322e706f6f6c2e663836333937332e6d30005bdc205a212c4573d32375b2f93ad215c21a34508ef89123a3ac5f8492552fd77c3d0fe7407ec6046268ed415600000000120000006e692e706f6f6c2e663836333937332e6d3000e8a88ee26c0c29b07da2ed2bea825cf08acbb57ee98f534aa7473eb27faa5d759127e33dff0810d01de24a9f5500000000130000007032702e706f6f6c2e663836333937332e6d3000373dca6a91c07ad49f86a8cac22b591eccec525bbe91ab11515b796a335ea67f77262491a45b3ef0c3621af032000000001a000000666f756e6472797573612e706f6f6c2e663836333937332e6d3000885ad5cbf24be2f85e01da691f16a0beb7d012fdeeb1a1d2f7efa8a14a48dabaf9af3a3877445b9061bfc31a17000000001a00000063686f7275732d6f6e652e706f6f6c2e663836333937332e6d30004f395f192fe94ea65559accb20bcf144700d3554a60553de67f054805839775946eb18db88c7bbe32f37b8121600000000190000006c756e616e6f7661322e706f6f6c2e663836333937332e6d30007b74680ad4633707d819b83f542232bcc88e66a28f23e6b0bf0d0c85d63502f8dfcc4709f5131fa2054d96c514000000001f00000070617468726f636b6e6574776f726b2e706f6f6c2e663836333937332e6d3000a7891814823af0fce30a3e5037534803fed58c0ab7e0de2aea9e8ef5982c3e31e9f6b1e32cf20e64cbcf644a0e000000001b0000007374616b657373746f6e652e706f6f6c2e663836333937332e6d300026366f84610fdd95923c16f9575e609ac5537adc7d5beb7b472ee7fea4db0379bf59b5c628e8f335029965ea0c00000000180000006c6561646e6f64652e706f6f6c2e663836333937332e6d3000acc27128aea3731cad4f045b0d011892f85857867a89305393b9bce796a316d5b4dd0ca1b6a3c32d64cf95450c000000001a000000626c6f636b73636f70652e706f6f6c2e663836333937332e6d30004eed9f8ec11536f0c3c8f5171365b2524afb37d07a2052e433654a2c514e3e39c1965f977a33cfb9106664250c000000001600000073687572696b2e706f6f6c2e663836333937332e6d300085872e16b7806fe2b385d4fc65609a68caade203edbf2f147209a1052ac3676b927431fdd83190f28cf6e0a90b0000000018000000647372766c6162732e706f6f6c2e663836333937332e6d30004a75120cf200659da75ee453d629fc0a411e5764652e441a3119d67097e20a983552a13ffb89479d704859860b0000000015000000616c3363352e706f6f6c2e663836333937332e6d3000a08174cb2460a626fbcca253e279ad175f02e12b7b358f07c10aacbb921b8e7aa7e564e77eea94faf4b063840b000000001b00000062617a696c696b6e6561722e706f6f6c2e663836333937332e6d30007d2b00c3ab4afbf5c2fd07372408a0130980e29621e70361e86bb50e768f25f5ff5219dad338575e1fc079610b000000001e000000626173696c69736b2d7374616b652e706f6f6c2e663836333937332e6d3000a73adabf811046d06cb097d6d6d967af363bf9485b8f3c63f968522ea2ff7b75e1fae5f09d0684313f5b51710a00000000140000006e306f6b2e706f6f6c2e663836333937332e6d3000b3a60b405427fe5267c85f12516339dce23c2d299781f17d6f7d4791edf75fb9a4bb0113fa4d495877eb008b08000000001e00000063617375616c706f6f6c746573742e706f6f6c2e663836333937332e6d300009e30367f2af386be37cd8786f4ba04dfa9742364eb634ae7512d402479972e0697d21afc5ead22f02d46c4a0500000000140000006b696c6e2e706f6f6c2e663836333937332e6d3000a0e96a90b0e21954c40e5e2fc7d56c412f1d52af0a8ea453d088a9c3659b8abd5c8e3d09e512ddf90ece274705000000001700000070726f706865742e706f6f6c2e663836333937332e6d3000f5bf6386a31f47147c5b4c6ba47b6948a24c10e56d3f57f5168f172a474dba6a38a48314e26cd3819e809e83040000000019000000626565317374616b652e706f6f6c2e663836333937332e6d300096360a2d46e98ba26fe18f88cfba3f225abe1986e413b7862d74c618e52082a5dbe360146bb088816b3bbf2301000000001b00000067657474696e676e6561722e706f6f6c2e663836333937332e6d30004193e2cd0630b71f1af526515f63863aad964e08eb451d2dd9454ec6041a4e7d35a53b0d3692875f81c91e1b010000000016000000617a657473692e706f6f6c2e663836333937332e6d3000140b7d9132eda1db94b9b6d30fc82d65b926ce4e3266457ff14dfe2b1a0525fe092655ade77d790b57c4d6b400000000001a000000737373757279616e73682e706f6f6c2e663836333937332e6d3000ad776676e1e8d4204591ac62dc61e453feeb6e2ad8d2b43d9e3c14b97241836a54d11b3714f1c8aeedabf3b100000000001700000074616e64656d6b2e706f6f6c2e663836333937332e6d300076d3435e83fe0cc8a6bcf1db156b4ff6231e9b8059ba17fb4c5fd779ae173d897debf4676ce4632ba41e20a900000000001500000031696e63342e706f6f6c2e663836333937332e6d300010fd8d4b9b3299315c1ea3c2b25d8c20fcc6bf45c080a355bf9f4723f24b50eac3bec18ad5c4fff11a2a08a4000000000014000000686e6e742e706f6f6c2e663836333937332e6d30009086e6fd0aa67b97242f668671adc7f167a32aa1514ce834a6c054b01b3078b694ecf65802cfb4f1df41169b000000000017000000677275626572782e706f6f6c2e663836333937332e6d30007b2c2b2286f93faa7cf0ed7963c82bab7d625b6a758369580f9bb952099c4d342821ed1162d63780cf8885970000000000150000006d6f6e646c696368742e706f6f6c2e6465766e657400aebdd9da72e4fe7cb226c712c19ad3c584cfe8c8c08dba1dbc753d3b06c448a99ecb91015725b05fa223398900000000001400000062672d302e706f6f6c2e663836333937332e6d30009f99bca8fd4c126fc1f9ade6ca38c0b308dfa8924e068a3666bb507b8764fc255447d109cf2cb28c922bcc8800000000001400000062672d312e706f6f6c2e663836333937332e6d30009c283a0348f37eab35769b210712958f3a6ce63aae5bb8f33f3e0167c6841d48aa37b11264b70bc0d5e8e9840000000000180000006a7374616b696e672e706f6f6c2e663836333937332e6d300009f7aaacc1dedaf1ccbe4402ed3d78458f3e39ccce3993c561d85f1312a718e3375385bf740c5dfb8b9bc0840000000000160000006267706e74782e706f6f6c2e663836333937332e6d3000bb5052c67b1cfcd47bd9610ab4b42b7895f7c7525d67e6848cd4dfcf6045dfecc63e4a490fe2b96ee6949981000000002e00000001005b9eaf36acd60cdb908b632926207b4beabc757053d30dceaf2d38fa8cc8dd5b66aaa6c7ec5793d02f5dfe4acc80cfcae93937bbb82195bb1cf2a27c2345e30b0100ef18e5860c7bcdcded19ccad492e56cbcf0fa18dd629d4ae1d0aeca6a0d3e9d76ecfbb701960612ffc2af4b677ecbedff82b05ee5131f28a466147c027e4ba0201003bd2567b0f11e2f913e006d5b8edde332cf7705eed4f035bad75a64d318517f9ee3ba2fbc4dfb8abadc7e40ba3f4f4cf07ee9ce62e2ef83e070207efbe191c060100652bb8009658ca9b570039c7d1a85e412c23cf33bbb242355a6c30279b379d030710b450b520b3b2eece73b170fa86b13583ce49dd1657dad555120fce68a80b000100b9252cc55c802222c70d74a3613529ae664df851a19f46e9965c0f42cb0c65ad6554e1ece8c6c8946aac014f39367922c230cfb0ae6e4877987161429fb31d0200010042f33e30d0331eac7ad912632629c0a3c9f8e705a297a6f6c399bf18c9223dbaae695eada75f09bc89231acc8dae6a8c2c56b91daaa908757c292e7ef8746e0e01004f330e40a0a730dec43cc04d6ca1d79e2cc74fb9cf30de0743466fbd36f663fcd6c48ae2c40d0b65f9bdbd8177fd8ae12e0cc7e812a81bff83e5562026e5110b01000b057a564005623775bbbb5752330c44e154d8ec8211821b8569097d647cc17122215c0dd5c59e5475f07457cd28f30f3a4c1c3769b67355bccf830f3f1de4060100d662f85a40b132113b161eca7b9edd6d0255de22dfacb5e78f7aa7980eb3f1567cfa6ef2d35e886f360d288145528e4b173deb34506520058830d4f17cefef05010071bc0b0774ab88ded4b1aa8eb493975aef7ec930c8b466af0e0f073956898ebeb4be3737fbf36e533654658f4d7aaf11f639da48f5db321308ae5a90ab783204010024477243abd763512950e11b7668dea1ad35e13350b0eb3548cb278708fe286f915919a028cfefa81279292f0332c8854ad08eb3fc75d17d098635dea683030601004d6724ccc00b70a1fe7de26483e670fe3707b843cdb5165366bbd250e9f68d3366c59b6d93353eaf04b17bbde4380ceb2c7887d2d422962f2220ddbaf4b9db0800010009018e4739222cfed6beb8c4cd0866e9cfd07083493fbc337f43b2f572a5d5f559b875f52941ebbcecae1f31a445f6c252b49f3cf398c137ba1d867698b2c4010000000001002949091c1498efe66c829b0364ff4b5e09c23694ae518f0a48d4602183b60c7338d8568c3e21ee343296b78d62635d895e4d1ce5789b4de5322c9c8be8154e0701008ce3a491229c31f88a673668512a143a2fec38815296e532250b7f5700b0edab6ed00d58e665c37ea0ef18487a5a72d49aaba3b90e8432a04762c5a7ffd6f4090001004a5db80b033c7f7c7ba9eced2da7b7ac975eeff0916f847252c24c905c0e528707c0b0bd8854dfae09c1a6eee15f7875bd3e04472345c560d6962cbc144ab8050100f9ed3f121970ecee4044b9e77d226e022a7c1d7a7d5b9df8afa6ea4a2621f18c76fb48205787b5c96de66b33141e426a72906be623b3f37064c4a1cda8f2f2080001008fd9feed12305aa408afad0e768464883bfb4fed486dc8005fda82fae30b2f85e1eadae6b3276a813d8d8efe6cf9d9680d2550338c9fb2cec86e1e56073bdf0a0100dd986496bfd6aaecdc4fe6b68927986d74ef9906eb28264d8881045e919043f10a88a9a6dbcbcc89314f03c5abc153aadf145c50646bdb6af86f58b0d38a4b050000010099254526beb115f80fb41796e0bfdc4746edcd0a28c79baa9b0241481c8eae7d6c748ba9975596e31f1f6a28b86b0f787c77c2486d6b5112380807bf2b39580800000100448d4253689dd905607c9bf5473b37f62558a53465b4fc319e579d0094a5eb9a2e13de3be079461fea62c875020a56577278189c81da2d7ca06df15deb38c10701002e795de1c0d3911070f6ca65367d7094cea4ac6a8fc5df91eb1c92cc8a1a223fa04456fad447c8ee2bfedd53157cdec36b222e1eadf491a385fbec49be448f0f01003ee11e60d9896719d868e69bcfbda1694108d5466e350d5ae1cdfc56ba10e052673ab2f6347bbd1a5898b786bfd037d9c3fbcd1119360a5ec4de067f348e44050100e95b682b215769b57a138a447ddc3965972c332d6e146435d504e6db5f96f96888bedf375f3d0f22c92429e350e543c40166de83aaa483e3030f07b0a2859601010078472fcc7367c745a70545cc0528676b65fd1ecd0f88c9f5a335168f577b5e175ffb36a29e741b0c1043a26cf0170dc78f70f012a4e4454f45590dc258164e060001007406657238d2051c867de61c23659cad959d6af505aa5069c4ab484ee1dd87deda54359520fc42e2a0c94b0b12d0faab5f8798eea862864e38c1ff8da028a90e0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000f1d0100000058afbe4d647c7899a757aef1bb39ab89ec6621214cd5d07108d3f180f139f1c30032982c167b54dfa0c6915b93635dd302b3f44f60003de0a2bea4a378d3db367a97275e9ddfe589c1137fc514b551a93e55d4de74b907ed43b23d4b48667c0fd7020000009006000073776170206f75743a207b2266726f6d5f636861696e223a2235353636383138353739363331383333303839222c22746f5f636861696e223a223830303031222c226f726465725f6964223a22307834383736663333373338623639376135663563343435366262643664623236653061306332376361373564313464373266653865306331343837396335316637222c22746f6b656e223a5b3131372c3131352c3130302c39392c34362c3130392c39372c3131322c34382c34382c35352c34362c3131362c3130312c3131352c3131362c3131302c3130312c3131365d2c2266726f6d223a5b3132302c3132312c3130382c3130352c34362c3131362c3130312c3131352c3131362c3131302c3130312c3131365d2c22746f223a5b3134302c3135352c36302c3137352c3132352c3233372c3231312c302c36332c38332c34392c33392c3132312c3139332c3138352c34332c3136312c39382c39332c3134385d2c22616d6f756e74223a2238363333353332222c22737761705f64617461223a2230783030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030363030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030313830303030303030303030303030303030303030303030303030343234643362636463393666343261633931396632373664376434663663393466323465303730333030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303130303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303230303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030366461323730323464643936303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303663383963363130303435343830303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303830303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030343030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303230303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030313431653031636634353033383038666233306631373830363033356138376366356135323137373237303030303030303030303030303030303030303030303030222c227261775f737761705f64617461223a7b22737761705f706172616d223a5b7b22616d6f756e745f696e223a2237393030303030303030303030303030303030222c226d696e5f616d6f756e745f6f7574223a2237383231303030303030303030303030303030222c2270617468223a2230783030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030323030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030222c22726f757465725f696e646578223a2230227d5d2c227461726765745f746f6b656e223a22307831653031636634353033383038666233306631373830363033356138376366356135323137373237222c226d61705f7461726765745f746f6b656e223a22307834323464336263646339366634326163393139663237366437643466366339346632346530373033227d7d9a040000636131636638636562663838343939343239636361386638376362636131356162386461666430363730323235396135333434646463653839656633663361356638363838383464343135303031303030303030303138333031333838316130343837366633333733386236393761356635633434353662626436646232366530613063323763613735643134643732666538653063313438373963353166373933373537333634363332653664363137303330333033373265373436353733373436653635373438633738373936633639326537343635373337343665363537343934386339623363616637646564643330303366353333313237373963316239326261313632356439343833383362636263623930316330303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303036303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303031383030303030303030303030303030303030303030303030303034323464336263646339366634326163393139663237366437643466366339346632346530373033303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030313030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030323030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303036646132373032346464393630303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030366338396336313030343534383030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030383030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303034303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030323030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303031343165303163663435303338303866623330663137383036303335613837636635613532313737323730303030303030303030303030303030303030303030303001000000f0d4fdbf6d9060bab8014a4eb10b8aae4527b70651569461a5e523292bb92e5c2ba754d82905000000cb133a6a77cfc61e00000000000000120000006d6f732e6d61703030372e746573746e65740203000000223022020000003b094f6018df5c924a985a963b47ed29c260d1ff9fb650df517e5247a749f930002eeb74a6177f588d80c0c752b99556902ddf9682d0b906f5aa2adbaf8466a4e900aa013b9336c728949c9e88bbc1aa761a8dbf91cd9d197d5eface6bbcd5a4b6faedf709a18935fad3ec33dc1bf0fa6397e6c2cbd2dc7b83d72f683cb30e50f920d966780600000000a62c1450d662861991aff6bd35555ad3854fc937476e411a1f1ea3a25ce259664ef4d14151cde795dfbc6eeda9998f3f4a51bbf46cabbdf1e58ceb4a4bbacdef61bc6f40c40c6dd4eeee04c6ab32b07b38b93615254309854653a06c75ff23b132a0f5a6e01f0874f11bfe3e088f8701d7a109e214d565196f574dc19cf0b863fd9c5e55455f3017f34b364e97d53023bc0c099d2d7c8fd7d400f9811516a034696dc081099ec36e62ebd84083d01c4c9660e027ee37b4776c03c2ab2d8b5de708e56760662244f90e000000aa013b9336c728949c9e88bbc1aa761a8dbf91cd9d197d5eface6bbcd5a4b6fa0088ea89f41509abfef35bbf43cedfb4d7d4db210ed829b73476c3e6cb474f4ca3019ef51562671539f335babcbeffbbe1e386e58da64131ffb4d7ebc5cb19bdfa8c00fa6a76539888989daa13d909654e0df40ef811a1ad713fbbf50f91d15fae793b003d0d2fdf53f91b4907a564f4c99a4b3ff4f903c26b63136afbffc4d846b6e24400c2d9f83be43033bc6b6b523ec4fe881b9fe8173cce72d7f9947da38d22acd8f1013ab4f993aeb1b313c5d2c671c9c7eb2ac41cac30d4fabab8dadd613575dababc0014cc31b717ccba7f508ae80ad7696ef70b63c289da484609fa1e8d1fa3c2ca7d007f29b983494e3513ea3710bc43cbd35032e9305ff3bade3866eed7dee24c94fc0095073f009a661b56d28b2ca843b30aae1aff033fcc2e01abc046228305f750bd004ae8f2cc0c13ad58634ee9718cfd06911a845e93d78adfa1b9b585c2305e8d0800cef92486ef4dd91bebe4dc27345e85ae8a82165f673181981704d1bc355ad658009fc14f87e6739602ad0e5f1211fcb9e0d59826b0d35d326d468842b73aacf66a005a8ca7fe98ead701132822284b29306097b3506253c07911a1450eca69ad727100000000'
-const eth2PolygonProof = '0xf9041af89b9486cbe739888bfbc0bca6e2d6106cfc5b3b1f69a5f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa00000000000000000000000009f477490aac940ce48249d8c455d8f6ae6dc29c0a0000000000000000000000000ae194f5d43ccab94c6dea994ff19001708d0e2eba00000000000000000000000000000000000000000000000000de0b6b3a7640000f89b9486cbe739888bfbc0bca6e2d6106cfc5b3b1f69a5f863a08c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925a00000000000000000000000009f477490aac940ce48249d8c455d8f6ae6dc29c0a0000000000000000000000000ae194f5d43ccab94c6dea994ff19001708d0e2eba0ffffffffffffffffffffffffffffffffffffffffffffffffe43e9298b137fffff902dd94ae194f5d43ccab94c6dea994ff19001708d0e2ebf863a0ca1cf8cebf88499429cca8f87cbca15ab8dafd06702259a5344ddce89ef3f3a5a00000000000000000000000000000000000000000000000000000000000000005a00000000000000000000000000000000000000000000000000000000000013881b902602ca61663b76d42c61157b026b970ac03e35da5d2da6bd5fe05268f07cf1f592000000000000000000000000000000000000000000000000000000000000000c0000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001400000000000000000000000000000000000000000000000000de0b6b3a76400000000000000000000000000000000000000000000000000000000000000000180000000000000000000000000000000000000000000000000000000000000001486cbe739888bfbc0bca6e2d6106cfc5b3b1f69a500000000000000000000000000000000000000000000000000000000000000000000000000000000000000149f477490aac940ce48249d8c455d8f6ae6dc29c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000149f477490aac940ce48249d8c455d8f6ae6dc29c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000c000000000000000000000000000000000000000000000000000000000000000600000000000000000000000000000000000000000000000000000000000000080000000000000000000000000424d3bcdc96f42ac919f276d7d4f6c94f24e0703000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000141e01cf4503808fb30f17806035a87cf5a5217727000000000000000000000000'
-module.exports = async (taskArgs) => {
-    const accounts = await ethers.getSigners()
-    const deployer = accounts[0];
-    const chainId = await deployer.getChainId();
-    let receiptProof = eth2PolygonProof;
-    // if (chainId === 97) {
-    //     receiptProof = bscReceiptProof;
-    // } else if (chainId === 80001) {
-    //     receiptProof = maticReceiptProof;
-    // } else {
-    //     throw new Error("chain id not supported", chainId)
-    // }
-    console.log("deployer address:",deployer.address);
-    let proxy = await hre.deployments.get("MAPOmnichainServiceProxyV2");
-    let mos = await ethers.getContractAt('MAPOmnichainServiceRelayV2',proxy.address);
-
-    await (await mos.connect(deployer).swapIn(
-        "5",
-        receiptProof,
-        {gasLimit: '10000000'}
-    )).wait();
-
-    console.log(`swap in`);
-}
